LayoutTests/http/tests/xmlhttprequest/xmlhttprequest-unsafe-redirect-expected.txt

This tests that unsafe redirects won't be allowed when making an XMLHttpRequest.
Sync XHR started.
readyState change 1
Error event.
Exception: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Async XHR started.
readyState change 1
SUCCESS: Error handler was called with readyState 1
